Nesting navigators
The Nesting navigators documentation was followed, in addition to separating the drawer in another file to make the code more readable. this fixes the problem of not having a correct flow of the navigator by the application
