+++ src/navigation/MainStack.js
@@ -1,31 +1,15 @@
 /* ----- library ----- */
 import React, {useEffect, useState} from 'react';
-import {Text, View, Image} from 'react-native';
-import {
-  createDrawerNavigator,
-  DrawerContentScrollView,
-} from '@react-navigation/drawer';
-import {subscriberUserId} from '../auth/cloudFirestore';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import {NavigationContainer} from '@react-navigation/native';
 import auth from '@react-native-firebase/auth';
-import {logout} from '../auth/authFirestore';
-import {useUser} from '../utils/user';
+import {Drawer} from './DrawerNavigation';
 /* ----- screens ----- */
-import Home from '../screens/Home';
-import Cuenta from '../screens/Cuenta';
 import Carrito from '../screens/Carrito';
-import Compras from '../screens/Compras';
 import Products from '../screens/Products';
 import Register from '../screens/Register';
-import Favoritos from '../screens/Favoritos';
-import MyProducts from '../screens/MyProducts';
 import ProductDetails from '../screens/ProductDetails';
-import Category from '../screens/Category';
-/* ----- components and styles ----- */
-import MenuStyles from './MenuStyles';
-import MenuButtom from '../components/atoms/MenuButtom';
-
+//https://reactnavigation.org/docs/nesting-navigators
 const MainStack = () => {
   const [user, setUser] = useState();
   const onAuthStateChanged = userStateChanged => setUser(userStateChanged);
@@ -33,124 +17,31 @@ const MainStack = () => {
     const subscriber = auth().onAuthStateChanged(onAuthStateChanged);
     return subscriber;
   }, []);
-  return user ? Drawer() : Login();
-};
-
-const DrawerStack = createDrawerNavigator();
-const Drawer = () => {
-  return (
-    <NavigationContainer>
-      <DrawerStack.Navigator
-        initialRouteName={'Home'}
-        screenOptions={{swipeEnabled: false, headerShown: false}}
-        drawerContent={props => MenuItems({...props})}>
-        <DrawerStack.Screen name="Home" component={Home} />
-        <DrawerStack.Screen name="Cart" component={Carrito} />
-        <DrawerStack.Screen name="Manage" component={MyProducts} />
-        <DrawerStack.Screen name="Account" component={Cuenta} />
-        <DrawerStack.Screen name="Products" component={Products} />
-        <DrawerStack.Screen name="Shopping" component={Compras} />
-        <DrawerStack.Screen name="WishList" component={Favoritos} />
-        <DrawerStack.Screen name="ProductDetails" component={ProductDetails} />
-        <DrawerStack.Screen name="Category" component={Category} />
-      </DrawerStack.Navigator>
-    </NavigationContainer>
-  );
+  return user ? App() : Login();
 };
 
 const LoginStack = createNativeStackNavigator();
 const Login = () => {
   return (
     <NavigationContainer>
-      <LoginStack.Navigator>
-        <LoginStack.Screen
-          name="Register"
-          component={Register}
-          options={{headerShown: false}}
-        />
+      <LoginStack.Navigator screenOptions={{headerShown: false}}>
+        <LoginStack.Screen name="Register" component={Register} />
       </LoginStack.Navigator>
     </NavigationContainer>
   );
 };
 
-const MenuItems = ({navigation}) => {
-  const user = useUser(state => state.user);
-  const storeUser = useUser(state => state.storeUser);
-  let {image, userName, userType, uid} = user;
-  let ImageUser =
-    'https://www.freeiconspng.com/thumbs/retail-store-icon/retail-store-icon-6.png';
-  if (image) {
-    ImageUser = image;
-  }
-  useEffect(() => {
-    const subscriber = subscriberUserId(uid, storeUser);
-    return () => subscriber();
-  }, [uid, storeUser]);
+const AppStack = createNativeStackNavigator();
+const App = () => {
   return (
-    <DrawerContentScrollView style={MenuStyles.container}>
-      <View style={MenuStyles.header}>
-        <Image
-          style={MenuStyles.Image}
-          source={{
-            uri: ImageUser,
-          }}
-        />
-        <View style={MenuStyles.columna}>
-          <Text style={MenuStyles.title}> Bright Shop </Text>
-          <Text style={MenuStyles.subtitle}>
-            {userName ? ' ' + userName : 'username'}
-          </Text>
-        </View>
-      </View>
-      <View style={MenuStyles.content}>
-        <MenuButtom
-          text="Home"
-          name="home"
-          onPress={() => navigation.navigate('Home')}
-        />
-        <MenuButtom
-          text="Shopping"
-          name="shopping-bag"
-          onPress={() => navigation.navigate('Shopping')}
-        />
-        <MenuButtom
-          text="WishList"
-          name="heart"
-          onPress={() => navigation.navigate('WishList')}
-        />
-        <MenuButtom
-          text="Cart"
-          name="shopping-cart"
-          onPress={() => navigation.navigate('Cart')}
-        />
-        {userType === 'Seller' && (
-          <>
-            <View style={MenuStyles.otheroptions}>
-              <Text style={MenuStyles.division}>Seller Options</Text>
-            </View>
-            <MenuButtom
-              text="New product"
-              name="plus"
-              onPress={() => navigation.navigate('Products')}
-            />
-            <MenuButtom
-              text="Manage my products"
-              name="gears"
-              onPress={() => navigation.navigate('Manage')}
-            />
-          </>
-        )}
-        <View style={MenuStyles.otheroptions}>
-          <Text style={MenuStyles.division}>Account Options</Text>
-        </View>
-        <MenuButtom
-          text="Account"
-          name="user"
-          onPress={() => navigation.navigate('Account')}
-        />
-        <MenuButtom text="Log Out" name="sign-out" onPress={logout} />
-      </View>
-    </DrawerContentScrollView>
+    <NavigationContainer>
+      <AppStack.Navigator screenOptions={{headerShown: false}}>
+        <AppStack.Screen name="Drawer" component={Drawer} />
+        <AppStack.Screen name="ProductDetails" component={ProductDetails} />
+        <AppStack.Screen name="Cart" component={Carrito} />
+        <AppStack.Screen name="Products" component={Products} />
+      </AppStack.Navigator>
+    </NavigationContainer>
   );
 };
